Fix missing dependencies in secure beacons
In some cases the _LdrLoadDll function would be given DllName's with mixed cases
and or different casing than what was listed in the `AllowedDll` array in main.h.
This change changes the strcmp call to _stricmp to ignore differences in casing
while loading dlls. This shouldn't make a difference in terms of security since
windows file paths are case insensitive by default.

Closes #20