Remove GetProcAddress -> LoadLibrary -> decrypt_string chains from BeaconRegisterC2
The GetProcAddress -> LoadLibrary -> decrypt_string chain in BeaconRegisterC2 results
in a memory and handle leak during beacon registration. `decrypt_string` returns a
malloc'ed pointer which is lost and never freed resulting in a memory leak. LoadLibrary
returns a HMODULE which is lost which means `FreeLibrary` cannot be used on that handle
to decrement the reference count. This means the module cannot be unloaded from the
beacons address space which may or may not be desirable. This change moves all of the
redundant LoadLibrary calls into one spot. Additionally it breaks the call chain up
in such a way `free` can be called on all of the decrypted_string chunks thus avoiding
a memory leak. This pattern can and probably should be abstracted into a function to
avoid repeated code.