Rework how ResBuffer is parsed
This change adjust how the ResBuffer was processed at the end of BeaconCallbackC2.
Previously the parent json object's reference count was never decremented preventing
it from being freed. This means that the string json buffer was also never freed
resulting in a memory leak. The parsing code now copies the data from the json object
string into an allocated buffer which is returned by the function. The json object
reference count is then decremented so the json object can be freed.