Remove free that was crashing beacon
While testing issue-33 changes I discovered the free on ResBuffer was
crashing the beacon. I still think that there may be a memory leak related
to how asprintf is being used on the ResBuffer however I am not sure and
it is better to have a small memory leak and a working beacon versus a
broken beacon.